Fixed instantiating
Now all instantiating of objects will be done in the given parent directory. For example the food spawners now has an additional field were the transform for carrot should be put so that the random generated carrots gets placed in the same place as the carrots that exist from the beginning.